BFloat16 conversion error fixed

+++ eval_util.py
@@ -67,7 +67,7 @@ def get_all_evals(cfg, model, tokenizer, eval_dataloader):
 
         probs = [p/v.item() for p, v in zip(probs, num_token_gt)]
 
-        eval_logs['gt_loss_per_token'] = eval_logs.get('gt_loss_per_token', []) + (gt_loss/num_token_gt).cpu().numpy().tolist()
+        eval_logs['gt_loss_per_token'] = eval_logs.get('gt_loss_per_token', []) + (gt_loss.float()/num_token_gt).cpu().numpy().tolist()
         eval_logs['gt_loss'] = eval_logs.get('gt_loss', []) + gt_loss.tolist()
         eval_logs['num_token_gt'] = eval_logs.get('num_token_gt', []) + num_token_gt.tolist()
         eval_logs['probs'] = eval_logs.get('probs', []) + probs
